Modifications to improve data loading speed on very large datasets. We now hold the raw data in arrays of bytes rather than shorts (hopefully reducing memory by 50%). The state table entries store both the raw string (eg A/T) and an array of the values. The string is used for comparisons while parsing the data which is significantly faster than array comparisons.

+++ src/flapjack/data/Line.java
@@ -29,7 +29,7 @@ public class Line
 	/**
 	 * Sets state information for a given loci position within a chromosome map.
 	 */
-	public void setLoci(int mapIndex, int lociIndex, short stateCode)
+	public void setLoci(int mapIndex, int lociIndex, int stateCode)
 	{
 		genotypes.get(mapIndex).setLoci(lociIndex, stateCode);
 	}
